backend/requirements.txt
# Notion Capture API - Backend Requirements

# Web Framework
fastapi==0.122.0
uvicorn==0.38.0
python-multipart==0.0.20

# Data Validation
pydantic==2.12.5
pydantic-settings>=2.0.0
pydantic_core==2.41.5

# Environment
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.24.0
requests>=2.31.0

# OpenAI
openai>=1.0.0

# Google APIs
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Image Processing / OCR
Pillow==10.4.0
pytesseract==0.3.10

# SSL Certificates
certifi>=2023.0.0

# Utils
typing_extensions==4.15.0
